'just_making_sure_everything_is_committed'

+++ server.py
@@ -101,7 +101,7 @@ def updatecountryByNamePage():
 	c['continent']=request.args.get('continent') 
 	c['area']=int(request.args.get('area'))
 	c['population']=int(request.args.get('population'))
-	c['gdp']=int(request.args.get('gdp'))
+	c['gdp']=float(request.args.get('gdp'))
 	return render_template(
 		'country.html',
 		c = c, alist=alist)
